Auto-generate external IDs in the database
Previously, the external ID was generated from the database primary key
in the application layer. This caused several problems:
- The ID was somewhat deterministic
- Open-sourcing the code would allow to reverse-engineer nobt-IDs, thus
making the whole encryption worthless.

Because we have to refresh the NobtEntity after persistence now (in
order to retrieve the generated value for external-ID), a transaction
MUST be open, otherwise em.flush and em.refresh fail. Thus, I had to
change the NobtRepositoryIT to use transactions via the CommandInvoker.