Add VirtualBeanPropertyWriter for Expense-links
The VirtualBeanPropertyWriter allows us to add dynamic properties every
time Jackson serializes an instance of our Expense. Through the built-in
attribute-passing mechanism of Jackson, we pass the PropertyWriter an
instance of our LinkFactory so that the logic for actually creating the
link lives outside of the PropertyWriter and can be easily tested.

We can now also use the "delete"-link in the ApiDocumentationTest to
delete the expense.