Add support for a scheme-overriding header
The new config value allows to define an HTTP-header that will is used
to override the scheme of generated links, if it contains a value. This
is needed for our production environment at AWS because our application
sits behind a load-balancer that accepts all requests forwards them to
our application. The load-balancer accepts HTTPS requests but forwards
them as HTTP. Thus, generated links would be invalid if we use the
scheme of the current request to generate the link.