Introduce 'minimal number of debts' optimizer
This new optimizer takes debts and tries to cover them through 'clever'
re-arrangement of existing debts. However, using only this one alone
results is not enough for a good outcome. That is why we also introduce
an OptimizerStrategyPipeline that allows us to apply multiple
strategies in a specific order. We use this to define the new
MINIMAL_AMOUNTS_AND_MINIMAL_NUMBER_OF_DEBTS optimizer which first
minimizes debts based on their amount and as a second step, minimizes
the number of debts.

MINIMAL_AMOUNTS_AND_MINIMAL_NUMBER_OF_DEBTS is the new default
optimizer for newly created nobts.