Add property-based consistency test
This test should ensure the most important of all properties:
The balance of a person, i.e. the amount they are getting paid/the
amount they have to pay MUST NOT change with an optimization.
The only thing that is allowed to change is who a person pays money to.